27/03/26 Afternoon

+++ rf/libs/sdr_HAL.c
@@ -17,11 +17,11 @@
  * @param target_freq Frecuencia deseada en Hz.
  * @param ppm_error Error de cristal en partes por millón.
  */
-static void tune_freq_with_ppm(hackrf_device* dev, uint64_t target_freq, int ppm_error) {
+static void tune_freq_with_ppm(hackrf_device* dev, uint64_t target_freq, float ppm_error) {
     double correction = 1.0 + ((double)ppm_error / 1000000.0);
     uint64_t corrected_freq = (uint64_t)((double)target_freq * correction);
     
-    printf("[HAL] Target: %lu Hz | PPM: %d | Tuning to: %lu Hz\n", 
+    printf("[HAL] Target: %lu Hz | PPM: %.3f | Tuning to: %lu Hz\n", 
            target_freq, ppm_error, corrected_freq);
 
     hackrf_set_freq(dev, corrected_freq);
